refactor: CardGameRound 타입을 constants로 이동하고 import 업데이트하기
