+++ frontend/src/contexts/CardGame/CardGameContext.ts
@@ -1,6 +1,7 @@
-import { CardGameRound, CardGameState, CardInfo } from '@/types/miniGame';
+import { CardGameState, CardInfo } from '@/types/miniGame';
 import { createContext, Dispatch, SetStateAction, useContext } from 'react';
 import { SelectedCardInfo } from './CardGameProvider';
+import { CardGameRound } from '@/constants/miniGame';
 
 type CardGameContextType = {
   startCardGame: boolean;
